Add docker configuration
This commit adds the Dockerfile file used for this project and a
docker-compose.yml file suitable for development purposes only.

The docker build isn't required to execute this project since the image
is pushed to the docker hub.

This also includes an example of env that will be used by the
application to be generated in the next interaction.