Adds the Transactions model
This model holds all information about accounts amount movement,
associating one Operation as a header for the transaction.

For a transfer is expected that two transactions are present and one
operation that connects these records. For a deposit operation, only one
transaction is expected.

We can't have any transaction with zero value, so we have a postgres
check at the database to ensure data consistency.