Adds TransferService
This service implements the money movement between accounts.

For now, this uses the BalanceService (inside a database lock) to check
if the source account has sufficient funds to ensure that we don't make
invalid calculations. Maybe this need to change later.

Only positive values can be used at this operation and the
implementation ensures that one transaction is created for each account
and associate then to the same Operation record.

This service need to ensures that the following invariant is always
respected:

  Operation.joins(:debit, :credit).sum("debits.amount + credits.amount")