Adds the controllers
This commit includes all controllers for operations defined. In all
cases we are just connecting the controller with the service or factory
that implements the operation necessary.

The POST /deposit returns the amount and have a custom JSON response.
All other endpoints returns the main record that are generated as JSON.

For errors, all endpoints must show the appropriate information.