Update project to use u-case gem
This PR adds the structure to migrate from old dry-* services
implementation to the new using u-case use cases.

At this point only the domain structure are created and in this first
implementation the use cases are using the services in the
implementation.

Some specs are added to the controllers to ensure that specs execute all
code from use cases and improve refactory safety. At this point use
cases specs is not present.